Move and remove types
