+++ src/components/ShowList/ShowList.tsx
@@ -2,7 +2,7 @@ import cn from 'classnames'
 import Link from 'next/link'
 import { FC } from 'react'
 
-import { ILinkList } from '../../types'
+import { ILinkList } from '../../types/dataLinks'
 
 import styles from './ShowList.module.scss'
