+++ src/types/dataLinks.ts
